Add a global volume control for audio input.

+++ SerialPrograms/Source/CommonFramework/AudioPipeline/IO/AudioSink.cpp
@@ -83,7 +83,7 @@ private:
 
 AudioSink::~AudioSink(){}
 
-AudioSink::AudioSink(Logger& logger, const AudioDeviceInfo& device, AudioChannelFormat format, float volume){
+AudioSink::AudioSink(Logger& logger, const AudioDeviceInfo& device, AudioChannelFormat format, double volume){
     NativeAudioInfo native_info = device.native_info();
     QAudioFormat native_format = native_info.preferredFormat();
